+++ packages/postcss/__tests__/integration.test.js
@@ -4,21 +4,12 @@ import { createConfig, match } from '@webpack-blocks/core'
 import postcss from '../index'
 
 test('Postcss works with defaults, without match()', t => {
-  const config = createConfig({}, [
-    postcss()
-  ])
+  const config = createConfig({}, [postcss()])
 
   t.deepEqual(config.module.rules, [
     {
       test: /\.css$/,
       use: [
-        'style-loader',
-        {
-          loader: 'css-loader',
-          options: {
-            minimize: undefined
-          }
-        },
         {
           loader: 'postcss-loader',
           options: {}
@@ -29,12 +20,7 @@ test('Postcss works with defaults, without match()', t => {
 })
 
 test('Postcss works with css() & match()', t => {
-  const config = createConfig({}, [
-    match('*.css', { exclude: /node_modules/ }, [
-      css(),
-      postcss()
-    ])
-  ])
+  const config = createConfig({}, [match('*.css', { exclude: /node_modules/ }, [css(), postcss()])])
 
   t.deepEqual(config.module.rules, [
     {
@@ -47,34 +33,7 @@ test('Postcss works with css() & match()', t => {
         },
         {
           loader: 'css-loader',
-          options: {
-            minimize: undefined
-          }
-        },
-        {
-          loader: 'postcss-loader',
           options: {}
-        }
-      ]
-    }
-  ])
-})
-
-test('Postcss should pass minimize option to css-loader', t => {
-  const config = createConfig({}, [
-    postcss({ minimize: true })
-  ])
-
-  t.deepEqual(config.module.rules, [
-    {
-      test: /\.css$/,
-      use: [
-        'style-loader',
-        {
-          loader: 'css-loader',
-          options: {
-            minimize: true
-          }
         },
         {
           loader: 'postcss-loader',
@@ -90,7 +49,7 @@ test('Postcss allows inline plugin config and custom options', t => {
 
   const config = createConfig({}, [
     postcss({
-      plugins: [ fakePostcssPlugin ],
+      plugins: [fakePostcssPlugin],
       parser: 'sugarss'
     })
   ])
@@ -99,17 +58,10 @@ test('Postcss allows inline plugin config and custom options', t => {
     {
       test: /\.css$/,
       use: [
-        'style-loader',
-        {
-          loader: 'css-loader',
-          options: {
-            minimize: undefined
-          }
-        },
         {
           loader: 'postcss-loader',
           options: {
-            plugins: [ fakePostcssPlugin ],
+            plugins: [fakePostcssPlugin],
             parser: 'sugarss'
           }
         }