+++ packages/sass/__tests__/integration.test.js
@@ -3,23 +3,13 @@ import { css } from '@webpack-blocks/assets'
 import { createConfig, match } from '@webpack-blocks/core'
 import sass from '../index'
 
-test('Sass works with defaults, without match()', t => {
-  const config = createConfig({}, [
-    sass()
-  ])
+test('Sass works with defaults, inside match()', t => {
+  const config = createConfig({}, [match('*.sass', [sass()])])
 
   t.deepEqual(config.module.rules, [
     {
-      test: /\.(sass|scss)$/,
+      test: /^.*\.sass$/,
       use: [
-        'style-loader',
-        {
-          loader: 'css-loader',
-          options: {
-            sourceMap: false,
-            minimize: undefined
-          }
-        },
         {
           loader: 'sass-loader',
           options: {}
@@ -29,13 +19,8 @@ test('Sass works with defaults, without match()', t => {
   ])
 })
 
-test('Sass works with css() & match()', t => {
-  const config = createConfig({}, [
-    match('*.sass', { exclude: /node_modules/ }, [
-      css(),
-      sass()
-    ])
-  ])
+test('Sass works with css()', t => {
+  const config = createConfig({}, [match('*.sass', { exclude: /node_modules/ }, [css(), sass()])])
 
   t.deepEqual(config.module.rules, [
     {
@@ -48,10 +33,7 @@ test('Sass works with css() & match()', t => {
         },
         {
           loader: 'css-loader',
-          options: {
-            sourceMap: false,
-            minimize: undefined
-          }
+          options: {}
         },
         {
           loader: 'sass-loader',
@@ -62,54 +44,18 @@ test('Sass works with css() & match()', t => {
   ])
 })
 
-test('Sass should pass sourceMap option to sass-loader and css-loader', t => {
-  const config = createConfig({}, [
-    sass({ sourceMap: true })
-  ])
+test('Sass should pass sourceMap option to sass-loader', t => {
+  const config = createConfig({}, [match('*.sass', [sass({ sourceMap: true })])])
 
   t.deepEqual(config.module.rules, [
     {
-      test: /\.(sass|scss)$/,
+      test: /^.*\.sass$/,
       use: [
-        'style-loader',
-        {
-          loader: 'css-loader',
-          options: {
-            sourceMap: true,
-            minimize: undefined
-          }
-        },
         {
           loader: 'sass-loader',
           options: {
             sourceMap: true
           }
-        }
-      ]
-    }
-  ])
-})
-
-test('Sass should pass minimize option to css-loader', t => {
-  const config = createConfig({}, [
-    sass({ minimize: true })
-  ])
-
-  t.deepEqual(config.module.rules, [
-    {
-      test: /\.(sass|scss)$/,
-      use: [
-        'style-loader',
-        {
-          loader: 'css-loader',
-          options: {
-            sourceMap: false,
-            minimize: true
-          }
-        },
-        {
-          loader: 'sass-loader',
-          options: {}
         }
       ]
     }