+++ packages/uglify/index.js
@@ -13,24 +13,28 @@ module.exports = uglify
  * @param {object} [options] UglifyJS options
  * @return {Function}
  */
-function uglify (options = {}) {
-  options = webpackMerge({
-    parallel: true,
-    cache: true,
-    uglifyOptions: {
-      compress: {
-        warnings: false
+function uglify(options = {}) {
+  options = webpackMerge(
+    {
+      parallel: true,
+      cache: true,
+      uglifyOptions: {
+        compress: {
+          warnings: false
+        }
       }
-    }
-  }, options)
+    },
+    options
+  )
 
   const postHook = (context, util) => {
     const plugin = new UglifyJSPlugin(options)
-    return util.addPlugin(plugin)
+    return util.merge({
+      optimization: {
+        minimizer: [plugin]
+      }
+    })
   }
 
-  return Object.assign(
-    () => prevConfig => prevConfig,
-    { post: postHook }
-  )
+  return Object.assign(() => prevConfig => prevConfig, { post: postHook })
 }