+++ packages/webpack/__e2e-fixtures__/babel-postcss-extract-text/webpack.config.js
@@ -4,32 +4,30 @@ const {
   entryPoint,
   match,
   performance,
+  setDevTool,
+  setMode,
   setOutput
 } = require('../../index')
 
 const babel = require('@webpack-blocks/babel')
 const postcss = require('@webpack-blocks/postcss')
 const extractText = require('@webpack-blocks/extract-text')
+const { css } = require('@webpack-blocks/assets')
 const path = require('path')
-const precss = require('precss')
+const postcssVars = require('postcss-simple-vars')
 
 module.exports = createConfig([
-  entryPoint(
-    path.join(__dirname, 'app.js')
-  ),
-  setOutput(
-    path.join(__dirname, 'build/bundle.js')
-  ),
+  setMode('development'),
+  setDevTool(false),
+  entryPoint(path.join(__dirname, 'app.js')),
+  setOutput(path.join(__dirname, 'build/bundle.js')),
   babel(),
   match('*.css', [
+    css(),
     postcss({
-      plugins: [
-        precss
-      ]
+      plugins: [postcssVars]
     }),
-    extractText(
-      'styles.css'
-    )
+    extractText('styles.css')
   ]),
   performance({
     maxAssetSize: 100000,