+++ packages/webpack/__e2e-fixtures__/postcss-sass-sourcemaps/webpack.config.js
@@ -1,21 +1,21 @@
-const { createConfig, entryPoint, setOutput } = require('../../index')
+const { createConfig, entryPoint, setMode, setOutput } = require('../../index')
 
 const path = require('path')
+const { match } = require('@webpack-blocks/webpack')
+const { css } = require('@webpack-blocks/assets')
 const postcss = require('@webpack-blocks/postcss')
-const precss = require('precss')
 const sass = require('@webpack-blocks/sass')
+const postcssVars = require('postcss-simple-vars')
 
 module.exports = createConfig([
-  entryPoint(
-    path.join(__dirname, 'styles.sass')
-  ),
-  setOutput(
-    path.join(__dirname, 'build/bundle.js')
-  ),
-  postcss({
-    plugins: [
-      precss
-    ]
-  }),
-  sass({ indentSyntax: true, sourceMap: true })
+  setMode('development'),
+  entryPoint(path.join(__dirname, 'styles.sass')),
+  setOutput(path.join(__dirname, 'build/bundle.js')),
+  match('*.sass', [
+    css({ sourceMap: true }),
+    postcss({
+      plugins: [postcssVars]
+    }),
+    sass({ indentSyntax: true, sourceMap: true })
+  ])
 ])