+++ packages/webpack/__e2e-fixtures__/typescript/webpack.config.js
@@ -1,19 +1,30 @@
-const { createConfig, defineConstants, entryPoint, performance, setOutput } = require('../../index')
+const {
+  createConfig,
+  defineConstants,
+  entryPoint,
+  performance,
+  setDevTool,
+  setMode,
+  setOutput
+} = require('../../index')
 
 const typescript = require('@webpack-blocks/typescript')
 const tslint = require('@webpack-blocks/tslint')
 const path = require('path')
 
 module.exports = createConfig([
-  entryPoint(
-    path.join(__dirname, 'app.ts')
-  ),
-  setOutput(
-    path.join(__dirname, 'build/bundle.js')
-  ),
-  typescript({ configFileName: path.resolve('./tsconfig.json') }),
+  setDevTool(false),
+  setMode('development'),
+  entryPoint(path.join(__dirname, 'app.ts')),
+  setOutput(path.join(__dirname, 'build/bundle.js')),
+  typescript({ configFileName: path.resolve(__dirname, './tsconfig.json') }),
   tslint({
-    emitErrors: true
+    emitErrors: true,
+    configuration: {
+      rules: {
+        'typedef-whitespace': false
+      }
+    }
   }),
   performance({
     maxAssetSize: 100000,