+++ packages/webpack/__tests__/end-to-end.test.js
@@ -1,12 +1,12 @@
 import test from 'ava'
 import fs from 'mz/fs'
-import jsdom from 'jsdom'
+import { JSDOM } from 'jsdom'
 import path from 'path'
 import webpack from 'webpack'
 
-const fixturesPath = path.join(__dirname, '..', '__e2e-fixtures__')
+const fixturesPath = path.join(__dirname, '../__e2e-fixtures__')
 
-test('building a minimal webpack project works', async (t) => {
+test('building a minimal webpack project works', async t => {
   const projectPath = path.join(fixturesPath, 'minimal')
   const buildPath = path.join(projectPath, 'build')
 
@@ -17,7 +17,7 @@ test('building a minimal webpack project works', async (t) => {
   t.is(bundleExports, 'I am the minimal test export')
 })
 
-test('building the babel/postcss/extract-text project works', async (t) => {
+test('building the babel/postcss/extract-text project works', async t => {
   const projectPath = path.join(fixturesPath, 'babel-postcss-extract-text')
   const buildPath = path.join(projectPath, 'build')
 
@@ -35,30 +35,32 @@ test('building the babel/postcss/extract-text project works', async (t) => {
   t.true(removeWhitespaces(styleContents).indexOf(removeWhitespaces('.app { margin: 40px; }')) > -1)
 })
 
-test('building the sass/extract-text project works', async (t) => {
+test('building the sass/extract-text project works', async t => {
   const projectPath = path.join(fixturesPath, 'sass-extract-text')
   const buildPath = path.join(projectPath, 'build')
 
   const config = require(path.join(projectPath, 'webpack.config.js'))
   await runWebpack(config)
 
-  global.window = await setUpJsdomEnv()
+  global.window = new JSDOM('<html><body></body></html>')
   global.document = global.window.document
   require(path.join(buildPath, 'bundle.js'))
 
   // Check if CSS file contains correct content
   const styleContents = await fs.readFile(path.join(buildPath, 'styles.css'), { encoding: 'utf8' })
-  t.true(removeWhitespaces(styleContents).indexOf(removeWhitespaces('body { padding: 25px; }')) > -1)
+  t.true(
+    removeWhitespaces(styleContents).indexOf(removeWhitespaces('body { padding: 25px; }')) > -1
+  )
 })
 
-test('building the typescript project works', async (t) => {
+test('building the typescript project works', async t => {
   const projectPath = path.join(fixturesPath, 'typescript')
   const buildPath = path.join(projectPath, 'build')
 
   const config = require(path.join(projectPath, 'webpack.config.js'))
   await runWebpack(config)
 
-  global.window = await setUpJsdomEnv()
+  global.window = new JSDOM('<html><body></body></html>')
   global.document = global.window.document
   require(path.join(buildPath, 'bundle.js'))
 
@@ -67,7 +69,7 @@ test('building the typescript project works', async (t) => {
   t.true(bundleContents.indexOf('module.exports = "This is the injected process.env.TEST!"') > -1)
 })
 
-test('the postcss/sass/source-maps project build does not fail', async (t) => {
+test('the postcss/sass/source-maps project build does not fail', async t => {
   // Regression test for https://github.com/andywer/webpack-blocks/issues/116
 
   const projectPath = path.join(fixturesPath, 'postcss-sass-sourcemaps')
@@ -78,44 +80,33 @@ test('the postcss/sass/source-maps project build does not fail', async (t) => {
   t.pass()
 })
 
-test('building the elm project works', async (t) => {
-  const projectPath = path.join(fixturesPath, 'elm')
-  const buildPath = path.join(projectPath, 'build')
-
-  const config = require(path.join(projectPath, 'webpack.config.js'))
-  await runWebpack(config)
-
-  global.window = await setUpJsdomEnv()
-  global.document = global.window.document
-  require(path.join(buildPath, 'bundle.js'))
-
-  t.pass()
-})
-
-test('building the sass/css-modules project works', async (t) => {
+test('building the sass/css-modules project works', async t => {
   const projectPath = path.join(fixturesPath, 'sass-css-modules')
   const buildPath = path.join(projectPath, 'build')
 
   const config = require(path.join(projectPath, 'webpack.config.js'))
   await runWebpack(config)
 
-  global.window = await setUpJsdomEnv()
+  global.window = new JSDOM('<html><body></body></html>')
   global.document = global.window.document
   require(path.join(buildPath, 'bundle.js'))
 
   // Check if CSS file contains correct content
   const styleContents = await fs.readFile(path.join(buildPath, 'styles.css'), { encoding: 'utf8' })
-  t.truthy(removeWhitespaces(styleContents).match(/\.styles--myClass--[0-9a-zA-Z]+\{margin:10px;\}/))
-  t.truthy(removeWhitespaces(styleContents).match(/\.styles--myClass--[0-9a-zA-Z]+:hover\{color:#ff0000;\}/))
+  t.truthy(removeWhitespaces(styleContents).match(/\.styles--myClass--[0-9a-zA-Z]+{margin:10px;}/))
+  t.truthy(
+    removeWhitespaces(styleContents).match(/\.styles--myClass--[0-9a-zA-Z]+:hover{color:#ff0000;}/)
+  )
 })
 
-function runWebpack (config) {
+function runWebpack(config) {
   return new Promise((resolve, reject) => {
     webpack(config, (error, stats) => {
       if (error) {
         reject(error)
       } else if (stats.hasErrors()) {
-        stats.toJson().errors.forEach((error) => console.error(error, '\n'))
+        // eslint-disable-next-line no-console
+        stats.toJson().errors.forEach(error => console.error(error, '\n'))
         reject(new Error('Webpack soft error occured. See stderr output.'))
       } else {
         resolve(stats)
@@ -124,18 +115,6 @@ function runWebpack (config) {
   })
 }
 
-function setUpJsdomEnv () {
-  return new Promise((resolve, reject) => {
-    jsdom.env('<html><body></body></html>', (error, window) => {
-      if (error) {
-        reject(error)
-      } else {
-        resolve(window)
-      }
-    })
-  })
-}
-
-function removeWhitespaces (string) {
+function removeWhitespaces(string) {
   return string.replace(/\s/g, '')
 }