+++ packages/assets/CHANGELOG.md
@@ -1,10 +1,23 @@
 # @webpack-blocks/assets - Changelog
 
+## 2.0.0-alpha.1
+
+- Major version upgrade of `css-loader` to the
+  [1.0.0](https://github.com/webpack-contrib/css-loader/releases/tag/v1.0.0)
+  ([#291](https://github.com/andywer/webpack-blocks/pull/291))
+- Major version upgrade of `file-loader` to the
+  [2.0.0](https://github.com/webpack-contrib/file-loader/releases/tag/v2.0.0)
+  ([#291](https://github.com/andywer/webpack-blocks/pull/291))
+
+## 2.0.0-alpha
+
+- Remove deprecated `fileType` API ([#260](https://github.com/andywer/webpack-blocks/issues/260))
+- Support for webpack 4 ([#261](https://github.com/andywer/webpack-blocks/pull/261))
+
 ## 1.0.0-rc
 
-- Added a `styleLoader` option to `css()` and `css.modules()` blocks.
-With it it's now possible to pass options to the `style-loader` or
-remove it from loaders completely.
+- Added a `styleLoader` option to `css()` and `css.modules()` blocks. With it it's now possible to
+  pass options to the `style-loader` or remove it from loaders completely.
 
 ## 1.0.0-beta.2